Morse Code T-SQL Solution
SOLUTION OVERVIEW:

This is a solution to the Morse Code problem using a T-SQL stored procedure on MS SQL Server 2008.

This stored procedure accepts a file path and an input file name as parameters. The file
contains multiple lines of Morse Codes. The output of the stored procedure is a result set
containing the English translation of the Morse Code lines, which can be used by any 
application code executing this stored procedure. Create this stored procedure in any 
database of choice. The DDL for the stored procedure usp_MorseDecipher can be found in the 
file named MorseDecipher.
	
Prerequisites:

1. On the SQL Server instance grant permission for 'Administer Bulk Operations'
	2. The SQL Server Service Agent must have read privilege on the folder where the input file resides
3. This stored procedure uses 4 files for processing. These files should be present in the
   same file path passed as the first parameter to this stored procedure. These files are:
	(a) The Morse Code input file (a sample provided in the packet
	(b) The Morse Code input format file called MorseInputFormat.fmt, which contains the metadata 
	    of the input file (included in the packet)
	(c) The Morse Code mapping file which contains the Morse Code to English alphabet mappings
	    (included in the packet)
	(d) The Morse Code mapping format file called MorseInputFormat.fmt, which contains the metadata 
	    of the mapping file (included in the packet)